Refactor PDF generation in GetBreachesPdfQueryHandler to use Handlebars for HTML templating and Puppeteer for PDF conversion; update Breach model to use long for PwnCount; add Handlebars.Net and PuppeteerSharp packages to project.

+++ BreachApi/Models/Breach.cs
@@ -9,7 +9,7 @@ public class Breach
     public DateTime AddedDate { get; set; }
     public DateTime ModifiedDate { get; set; }
     public string Description { get; set; } = string.Empty;
-    public int PwnCount { get; set; }
+    public long PwnCount { get; set; }
     public string LogoPath { get; set; } = string.Empty;
     // Add other fields as needed from the API
 }
